Commented out following hook:
 after 'deploy:publishing', 'deploy:restart'

because of following reason:

deploy:restart automatically tries to run unicorn server restart script:
For e.g. see the log below:
  # INFO [e00fcb66] Running /usr/bin/env sudo /etc/init.d/unicorn_cebucondo_production restart as cebucond@106.186.120.245
Thus if you application doesn't use Unicorn server then there is no need to restart it. How to change it to any other server that's a different story and not in scope
right now.

This change helped in making a successful deployment on the remote host.